Regras de negócio para transação

+++ src/main/resources/db/migration/V2__create_account_balance_table.sql
@@ -1,6 +1,6 @@
 -- Criação da tabela de saldo de conta
 CREATE TABLE account_balance (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id NUMERIC PRIMARY KEY AUTO_INCREMENT,
     cpf NUMERIC(11,0) PRIMARY KEY, -- A chave primária é o CPF
     saldo DOUBLE NOT NULL,
     FOREIGN KEY (cpf) REFERENCES pix_transaction(cpf)
